# docs/swagger.yaml
basePath: /api
definitions:
...
  database.Feature:
    properties:
      category:
        $ref: '#/definitions/pgtype.Text'
      created_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      description:
        $ref: '#/definitions/pgtype.Text'
      feature_code:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  handlers.CreateFeatureRequest:
    properties:
      category:
        example: analytics
        type: string
      description:
        example: Enables advanced analytics dashboards
        type: string
      feature_code:
        example: ADVANCED_ANALYTICS
        type: string
      name:
        example: Advanced Analytics
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.Timestamptz:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
host: localhost:3000
info:
  contact: {}
  description: Cloud License Platform — multi-tenant SaaS licensing and entitlement
    management.
  title: CLP API
  version: "1.0"
paths:
...
  /features:
    get:
      description: Returns the full feature catalog ordered by feature code.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Feature'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List all features
      tags:
      - features
    post:
      consumes:
      - application/json
      description: Adds a new feature to the catalog.
      parameters:
      - description: Feature
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateFeatureRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/database.Feature'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a feature
      tags:
      - features
  /features/{code}:
    get:
      description: Returns a feature by its unique feature code.
      parameters:
      - description: Feature code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Feature'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get feature by code
      tags:
      - features
  /features/{id}:
    delete:
      description: Permanently removes a feature from the catalog by UUID.
      parameters:
      - description: Feature UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete a feature
      tags:
      - features
...
swagger: "2.0"